fix: session restart message on heartbeat lapse
When the heartbeat for the connection gets lased there is a process for
restarting the active connection to see if it helps. When restarting the
active connection we try to close and open the browser which on a cloud
provider means creation of a new session, this inturn updates the
mapping of the browser id with the session id and thus when the process
is getting closed the previous old sessions remains unmarked on the user
dashboard.

We can basically mark the connection before closing the browser, with
the remark of restarting becasue, then user will be aware that the
session is restarted and won't be unmarked.

Ref: https://github.com/DevExpress/testcafe/pull/5535